Imported Upstream version 4.0.0-dfsg

+++ src/VBox/Runtime/common/rand/randadv.cpp
@@ -1,4 +1,4 @@
-/* $Id: randadv.cpp $ */
+/* $Id: randadv.cpp 34507 2010-11-30 13:14:14Z vboxsync $ */
 /** @file
  * IPRT - Random Numbers, Generic Glue.
  */
@@ -368,10 +368,10 @@ DECLCALLBACK(uint64_t)  rtRandAdvSynthesizeU64FromU32(PRTRANDINT pThis, uint64_t
 {
     uint64_t off = u64Last - u64First;
     if (off <= UINT32_MAX)
-        return (uint64_t)pThis->pfnGetU32(pThis, 0, off) + u64First;
+        return (uint64_t)pThis->pfnGetU32(pThis, 0, (uint32_t)off) + u64First;
 
     return (    (uint64_t)pThis->pfnGetU32(pThis, 0, UINT32_MAX)
-            |  ((uint64_t)pThis->pfnGetU32(pThis, 0, off >> 32) << 32))
+            |  ((uint64_t)pThis->pfnGetU32(pThis, 0, (uint32_t)(off >> 32)) << 32))
          + u64First;
 }
